+++ RELEASE_NOTES.md
@@ -9,6 +9,8 @@
 - **Data**: Add ```INTERNAL_ID``` and ```CONSOLIDATION_MODE``` columns for ```DIM_SNAPSHOTS``` table
 - **Data**: Add ```APP_VIOLATIONS_EVOLUTION``` table
 - **Data**: Add ```MOD_VIOLATIONS_EVOLUTION``` table
+- **Data**: Add ```APP_TECHNO_SIZING_MEASURES``` table
+- **Data**: Add ```MOD_TECHNO_SIZING_MEASURES``` table
 
 #### Bug Fixes